in branch, working on maximum filter for dilating histograms
Implimented Daniel Lemire's algorithm.  Trying to adapt it for a
centered window.


git-svn-id: https://subversion.int.janelia.org/myerslab/projects/whisker/branches/2009-05-25@373 fbdb9df1-db47-0410-955a-fa45b24b025c

+++ common.h
@@ -1,6 +1,8 @@
 #ifndef _H_COMMON
 #define _H_COMMON
 
+#include <stdlib.h>
+
 // #include <vld.h> // for VC++: Visual Leak Detector (v1.9b) 
                     // - comment out if this causes problems
                     // - comment in  to check for memory leaks in an
@@ -40,7 +42,12 @@
     ...
   }
 */
-void *request_storage( void *buffer, int *maxlen, int nbytes, int minindex, char *msg );
+void *request_storage( 
+    void   *buffer  , 
+    size_t *maxlen  , 
+    size_t  nbytes  , 
+    size_t  minindex, 
+    char   *msg );
 
 #endif //_H_COMMON
